requirements.txt
flask>=3.0
flask-sqlalchemy>=3.1
networkx>=3.2
numpy>=1.26
matplotlib>=3.8
Pillow>=10.0
psycopg2-binary>=2.9
requests>=2.31
python-dotenv>=1.0
